feat:task

+++ app/task/internal/repository/db/model/task.go
@@ -1,9 +1,9 @@
 package model
 
 type Task struct {
-	TaskID    uint `gorm:"primarykey"` // id
-	UserID    uint `gorm:"index"`      // 用户id
-	Status    int  `gorm:"default:0"`
+	TaskID    int64 `gorm:"primarykey"` // id
+	UserID    int64 `gorm:"index"`      // 用户id
+	Status    int   `gorm:"default:0"`
 	Title     string
 	Content   string `gorm:"type:longtext"`
 	StartTime int64
